refactor: FE-24 - padroniza logica de error notification em paineis operacionais
- Simplifica hasNotifiedError ref em OpsSummarySection, OpsEventsPanel e OpsDeliveryExceptionsPanel
- Corrige imports em dashboard/index.tsx para paths atualizados
- Atualiza backlog com FE-24 concluido
